+++ changelog.md
@@ -2,7 +2,14 @@
 
 All notable changes to `EtagConditionals` will be documented in this file.
 
-## Version 1.0
+## Version 1.0.3
+Rework of IfMatch internals + tests updated
 
+
+## Version 1.0.0
+Initial release 
 ### Added
-- Everything
+- IfMatch Middleware 
+- IfNoneMatch Middleware 
+- SetEtag Middleware 
+- Tests